+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Release history for s390-tools (MIT version)
 
   Bug Fixes:
   - dbginfo.sh: save dbginfo.sh version to dbginfo.log
+  - dbginfo.sh: ensure compatibility with /bin/dash shell
 
 * __v2.23.0 (2022-08-18)__